--- database/pgsqlIndex.js
@@ -1,7 +1,0 @@
-const { Client } = require('pg'); //switch from client to pool as an optimization later.
-
-const client = new Client({
-  database: 'images',
-  user: 'danmoy',
-  password: '',
-});
