+++ server/index.js
@@ -13,8 +13,7 @@ app.use(express.static('./public'));
 
 app.get('/carousel/:id', (req, res, next) => {
   let id = req.params.id;
-  Images.find({id: id})
-  .exec((err, result) => {
+  Images.getImages(id, (err, result) => {
     if (err) {
       console.log('error get request', err);
       res.sendStatus(404);
@@ -23,67 +22,6 @@ app.get('/carousel/:id', (req, res, next) => {
     }
   })
 });
-
-//write a post route
-app.post('/api/carousel', (req, res) => {
-  //pass array to a function that writes to the DB
-  const newImages = [];
-  for (var i = 0; i < req.body.length; i++) {
-    const newImage = new Images(req.body[i]);
-    newImages.push(newImage);
-  }
-  console.log(newImages);
-  Images.insertMany(newImages, (err, result) => {
-    if (err) {
-      console.log('error get request', err);
-      res.sendStatus(404);
-    } else {
-      res.json(result);
-    };
-  })
-});
-
-// //write a get route
-app.get('/api/carousel/:refId', (req, res) => {
-  //pass refId to a function that finds the entry with refId in the DB
-  let refId = req.params.refId;
-  Images.find({ _id: refId }, (err, result) => {
-    if (err) {
-      console.log('error get request', err);
-      res.sendStatus(404);
-    } else {
-      res.json(result);
-    };
-  })
-})
-
-// //write a put route
-app.put('/api/carousel/:refId', (req, res) => {
-  //pass req.body to a function that updates the entry at refId in the DB
-  let refId = req.params.refId;
-  Images.updateOne({ _id: refId }, req.body, (err, result) => {
-    if (err) {
-      console.log('error get request', err);
-      res.sendStatus(404);
-    } else {
-      res.json(result);
-    };
-  })
-});
-
-// //write a delete route
-app.delete('/api/carousel/:refId', (req, res) => {
-  //pass refId to a function that deletes the entry with refId in the DB
-  let refId = req.params.refId;
-  Images.deleteOne({ _id: refId }, (err, result) => {
-    if (err) {
-      console.log('error get request', err);
-      res.sendStatus(404);
-    } else {
-      res.json(result);
-    };
-  })
-})
 
 module.exports = app.listen(port, () => {
   console.log(`listening on port ${port}`);
